gov_analysis/src/collectors/un_egov_collector.py
import pandas as pd
import requests
from bs4 import BeautifulSoup
from typing import Dict, List, Optional
from datetime import datetime

class UNEGovCollector:
    def __init__(self):
        self.base_url = "https://publicadministration.un.org/egovkb/en-us/Data-Center"
    
    def fetch_egov_development_index(self) -> pd.DataFrame:
        """
        Fetch E-Government Development Index (EGDI) data
        EGDI is a composite measure of three dimensions:
        provision of online services, telecommunication connectivity,
        and human capacity
        """
        try:
            # Note: Since UN doesn't provide a direct API, we'll need to
            # implement web scraping or download data manually.
            # This is a placeholder for the actual implementation
            print("Note: UN E-Government data typically needs manual download from their website")
            return pd.DataFrame()
            
        except Exception as e:
            print(f"Error fetching UN EGDI data: {str(e)}")
            return pd.DataFrame()
    
    def fetch_online_service_index(self) -> pd.DataFrame:
        """
        Fetch Online Service Index (OSI) data
        Measures scope and quality of online services
        """
        try:
            # Placeholder for OSI data collection
            # Actual implementation would involve web scraping or API calls
            return pd.DataFrame()
            
        except Exception as e:
            print(f"Error fetching UN OSI data: {str(e)}")
            return pd.DataFrame()
    
    def save_data(self, df: pd.DataFrame, category: str):
        """Save collected data to CSV"""
        if not df.empty:
            timestamp = datetime.now().strftime("%Y%m%d")
            filename = f"data/raw/un_{category}_{timestamp}.csv"
            df.to_csv(filename, index=False)
            print(f"Saved UN {category} data to {filename}")
