gov_analysis/requirements.txt
pandas>=1.5.3
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
sodapy==2.2.0  # for accessing government open data
selenium==4.11.2
plotly>=5.16.1
numpy>=1.24.0
scikit-learn>=1.0.2
pymongo>=4.4.1  # for local data storage
